+++ x/evmutil/genesis_test.go
@@ -6,6 +6,7 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	sdkmath "cosmossdk.io/math"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	"github.com/kava-labs/kava/x/evmutil"
 	"github.com/kava-labs/kava/x/evmutil/testutil"
 	"github.com/kava-labs/kava/x/evmutil/types"
@@ -28,7 +29,7 @@ func (s *genesisTestSuite) TestInitGenesis_SetAccounts() {
 	)
 	accounts := s.Keeper.GetAllAccounts(s.Ctx)
 	s.Require().Len(accounts, 0)
-	evmutil.InitGenesis(s.Ctx, s.Keeper, gs)
+	evmutil.InitGenesis(s.Ctx, s.Keeper, gs, s.AccountKeeper)
 	accounts = s.Keeper.GetAllAccounts(s.Ctx)
 	s.Require().Len(accounts, 1)
 	account := s.Keeper.GetAccount(s.Ctx, s.Addrs[0])
@@ -47,7 +48,7 @@ func (s *genesisTestSuite) TestInitGenesis_SetParams() {
 		[]types.Account{},
 		params,
 	)
-	evmutil.InitGenesis(s.Ctx, s.Keeper, gs)
+	evmutil.InitGenesis(s.Ctx, s.Keeper, gs, s.AccountKeeper)
 	params = s.Keeper.GetParams(s.Ctx)
 	s.Require().Len(params.EnabledConversionPairs, 1)
 	s.Require().Equal(conversionPair, params.EnabledConversionPairs[0])
@@ -61,8 +62,23 @@ func (s *genesisTestSuite) TestInitGenesis_ValidateFail() {
 		types.DefaultParams(),
 	)
 	s.Require().Panics(func() {
-		evmutil.InitGenesis(s.Ctx, s.Keeper, gs)
+		evmutil.InitGenesis(s.Ctx, s.Keeper, gs, s.AccountKeeper)
 	})
+}
+
+func (s *genesisTestSuite) TestInitGenesis_ModuleAccount() {
+	gs := types.NewGenesisState(
+		[]types.Account{},
+		types.DefaultParams(),
+	)
+	s.Require().NotPanics(func() {
+		evmutil.InitGenesis(s.Ctx, s.Keeper, gs, s.AccountKeeper)
+	})
+	// check for module account this way b/c GetModuleAccount creates if not existing.
+	acc := s.AccountKeeper.GetAccount(s.Ctx, s.AccountKeeper.GetModuleAddress(types.ModuleName))
+	s.Require().NotNil(acc)
+	_, ok := acc.(authtypes.ModuleAccountI)
+	s.Require().True(ok)
 }
 
 func (s *genesisTestSuite) TestExportGenesis() {